ref: AIkido complete refactor
