+++ .env.example
@@ -1,0 +1,7 @@
+AIKIDO_HOST="0.0.0.0"
+AIKIDO_HTTP_PORT=80
+AIKIDO_HTTPS_PORT=443
+AIKIDO_ENABLE_HTTP=true
+AIKIDO_ENABLE_HTTPS=true
+AIKIDO_KEY_FILE='cert/server.key'
+AIKIDO_CERT_FILE='cert/server.pem'
