+++ requirements.txt
@@ -3,3 +3,4 @@ neotermcolor
 hypercorn
 quart
 requests
+python-dotenv
